+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [v0.2.8] - 2026-06-17
+
+Wired up search for the kit
+
+### Changed
+
+- Took the placeholder search icon and added simple search
+
 ## [v0.2.7] - 2026-06-17
 
 Wired up documentation in the kit